requirements.txt
Flask==1.1.1
requests==2.23.0
scikit-learn==0.22.2.post1
numpy==1.18.3
tqdm==4.38.0
torch==1.4.0
matplotlib~=3.2.1
pandas~=1.0.3
seaborn~=0.10.0
livelossplot~=0.5.0
tensorboardx
pytorch-crf
nltk==3.5
gensim==3.6.0